+++ code/hyperdb/src/helpers/db/main.ts
@@ -2,3 +2,4 @@ import * as ipc from './ipc.main';
 
 export { ipc };
 export { Db } from './Db.main';
+export { create } from './main.create';
